requirements.txt
streamlit>=1.28.0
openai>=1.12.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
google-auth-oauthlib>=1.1.0
httplib2>=0.22.0
certifi>=2023.7.22
faiss-cpu>=1.7.4
bcrypt>=4.0.1
google-api-python-client>=2.129.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
streamlit-authenticator==0.4.2




